+++ lib/index.d.ts
@@ -1,3 +1,0 @@
-export { NightProvider } from './context';
-export { useNight } from './useNight';
-export { withNight } from './withNight';