+++ lib/withNight.d.ts
@@ -1,2 +1,0 @@
-import React from 'react';
-export declare const withNight: (WrappedComponent: React.ComponentType<any>) => (props: any) => JSX.Element;