+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # react-night Changelog
 
 ## 1.0.0 alpha 0 (Soon)
+
 #### New Feature
-- `NightProvider` context API
+
+- `NightProvider` context
 - `useNight` hook
 - `withNight` HOC